kodama-go: permit the use of empty matrices in linkage
This commit fixes a corner case in the Go bindings for Kodama that seems
to arise more frequently than expected: running linkage on an empty
matrix. While callers could detect this edge case and work around it in
their own specific way, it makes sense for Kodama to be consistent in this
respect and permit a no-op linkage run on an empty matrix.

Fixing this was a bit subtle. Apparently, the idiomatic way to pass a
pointer to a slice from Go to C is to take a pointer to its first element
like so: &slice[0]. This makes sense. The problem with this approach is,
obviously, that it doesn't work if the slice is empty, which is exactly
the case we want to support here. In this case, we need to extract the
underlying pointer more explicitly.

Secondarily, we also need to be careful not to send a nil pointer to FFI
bindings. Namely, in a previous commit, we fixed some undefined behavior
by requiring a non-null pointer. Thus, the Go bindings must live up to that
contract.

There are other ways we could have fixed this. For example, we could have
detected this edge case in the Go code and avoid calling through to the
kodama library altogether---with some additional case analysis on the
public API of Dendrogram. But the Kodama library specifically supports this
edge case, and it is more elegant to avoid additional special case
analysis.