style: fix deprecated stuff
This fixes/removes stuff that has been deprecated.

The first change is to remove the 'description' and 'cause' methods on
the Error trait impl. Both of those methods have been deprecated. The
former in favor of a std::fmt::Display impl which we already have, and
the latter in favor of a new 'source' method. Since there is no
underlying error to return, we can omit it altogether.

The second change is to fix our use of trait objects, updating to the
new 'dyn' syntax. This is a strictly syntactical change.